change to ApartmentList and ApartmentDetail
