+++ src/pages/client/ApartmentList/ApartmentList.jsx
@@ -1,293 +1,430 @@
-import { useState } from "react";
-import styled from "styled-components";
-import { FaSearch, FaBed, FaBath, FaCar } from "react-icons/fa";
-import { BsHouseDoor } from "react-icons/bs";
-/* Import Link từ react-router-dom */
-import { Link } from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import { 
+  Card, 
+  Row, 
+  Col,
+  Avatar, 
+  Space, 
+  Tag, 
+  Input, 
+  Select, 
+  Button, 
+  Pagination,
+  Skeleton,
+  Typography,
+  Badge,
+  Tooltip,
+  Divider,
+  Image
+} from "antd";
+import { 
+  HomeOutlined, 
+  DollarOutlined, 
+  UserOutlined, 
+  SearchOutlined, 
+  FilterOutlined,
+  EnvironmentOutlined,
+  AreaChartOutlined,
+  EyeOutlined,
+  ArrowRightOutlined
+} from "@ant-design/icons";
+import { useNavigate } from "react-router-dom";
 
-const Container = styled.div`
-  background: #f3f4f6;
-  min-height: 100vh;
-`;
+const { Search } = Input;
+const { Option } = Select;
+const { Title, Text, Paragraph } = Typography;
 
-const Wrapper = styled.div`
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 16px;
-`;
-
-const TopBar = styled.div`
-  display: flex;
-  gap: 24px;
-  align-items: center;
-  background: white;
-  color: #dc2626;
-  padding: 8px 16px;
-  font-size: 14px;
-  font-weight: 500;
-`;
-
-const MainNavBar = styled.div`
-  display: flex;
-  align-items: center;
-  background: white;
-  padding: 16px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-`;
-
-const Logo = styled.h1`
-  margin: 0;
-  font-size: 24px;
-  color: #111827;
-`;
-
-const NavLinks = styled.div`
-  display: flex;
-  align-items: center;
-  margin-left: 24px; 
-  gap: 16px;
-  flex: 1; 
-`;
-
-const Spacer = styled.div`
-  flex: 1; 
-`;
-
-const NavItem = styled(Link)`
-  text-decoration: none;
-  font-weight: 500;
-  padding: 8px 16px;
-  border-radius: 4px;
-  color: #1f2937;
-  background: transparent;
-
-  &:hover {
-    background: #16a34a;
-    color: #fff;
+// Dữ liệu mẫu các bài đăng căn hộ
+const sampleApartments = [
+  {
+    id: 1,
+    title: "Căn hộ cao cấp Vinhomes Central Park",
+    description: "Căn hộ ban công rộng, view sông, nội thất cao cấp, an ninh 24/7",
+    address: "Quận Bình Thạnh, TP.HCM",
+    price: 5800000,
+    area: 70,
+    bedrooms: 2,
+    bathrooms: 2,
+    owner: "Văn Phú Invest",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=1",
+    date: "2025-03-12",
+    category: "Cho thuê",
+    tags: ["Cao cấp", "Đầy đủ nội thất", "View sông"],
+    views: 145,
+    status: "Còn trống",
+    image: "https://picsum.photos/400/300?random=1"
+  },
+  {
+    id: 2,
+    title: "Studio căn hộ The Sun Avenue",
+    description: "Căn hộ đầy đủ nội thất, cửa sổ lớn, bảo vệ 24/7, gần trung tâm thương mại",
+    address: "Quận 2, TP.HCM",
+    price: 3500000,
+    area: 35,
+    bedrooms: 1,
+    bathrooms: 1,
+    owner: "Landsoft Building",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=2",
+    date: "2025-03-10",
+    category: "Cho thuê",
+    tags: ["Studio", "Đầy đủ nội thất", "Gần trung tâm"],
+    views: 98,
+    status: "Đã cho thuê",
+    image: "https://picsum.photos/400/300?random=2"
+  },
+  {
+    id: 3,
+    title: "Căn hộ 3 phòng ngủ Masteri Thảo Điền",
+    description: "Căn góc, view đẹp, nội thất cao cấp, bảo vệ 24/7, hồ bơi, gym",
+    address: "Quận 2, TP.HCM",
+    price: 7200000,
+    area: 95,
+    bedrooms: 3,
+    bathrooms: 2,
+    owner: "Osaka Complex",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=3",
+    date: "2025-03-08",
+    category: "Cho thuê",
+    tags: ["Căn góc", "Nội thất cao cấp", "View đẹp"],
+    views: 167,
+    status: "Còn trống",
+    image: "https://picsum.photos/400/300?random=3"
+  },
+  {
+    id: 4,
+    title: "Penthouse Saigon Pearl",
+    description: "Penthouse sang trọng, view toàn thành phố, sân vườn riêng, an ninh tuyệt đối",
+    address: "Quận Bình Thạnh, TP.HCM",
+    price: 15000000,
+    area: 180,
+    bedrooms: 4,
+    bathrooms: 3,
+    owner: "Sky Garden",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=4",
+    date: "2025-03-05",
+    category: "Bán",
+    tags: ["Penthouse", "Sang trọng", "View toàn thành phố"],
+    views: 213,
+    status: "Đang đặt cọc",
+    image: "https://picsum.photos/400/300?random=4"
+  },
+  {
+    id: 5,
+    title: "Căn hộ 2 phòng ngủ Gateway Thảo Điền",
+    description: "Căn hộ thiết kế hiện đại, view sông, bảo vệ 24/7, gần trạm metro",
+    address: "Quận 2, TP.HCM",
+    price: 6000000,
+    area: 65,
+    bedrooms: 2,
+    bathrooms: 2,
+    owner: "Văn Phú Invest",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=5",
+    date: "2025-03-01",
+    category: "Cho thuê",
+    tags: ["Hiện đại", "View sông", "Gần metro"],
+    views: 187,
+    status: "Còn trống",
+    image: "https://picsum.photos/400/300?random=5"
+  },
+  {
+    id: 6,
+    title: "Căn hộ cao cấp Phú Đông",
+    description: "Căn hộ thiết kế sang trọng, đầy đủ tiện nghi, an ninh tốt, tiện ích đầy đủ",
+    address: "TP. Thủ Đức, TP.HCM",
+    price: 4500000,
+    area: 55,
+    bedrooms: 2,
+    bathrooms: 1,
+    owner: "Landsoft Building",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=6",
+    date: "2025-02-25",
+    category: "Cho thuê",
+    tags: ["Mới xây", "Hiện đại", "Tiện ích đầy đủ"],
+    views: 142,
+    status: "Còn trống",
+    image: "https://picsum.photos/400/300?random=6"
+  },
+  {
+    id: 7,
+    title: "Căn hộ dịch vụ quận 1",
+    description: "Căn hộ dịch vụ cao cấp, full nội thất, dịch vụ hàng ngày, vị trí trung tâm",
+    address: "Quận 1, TP.HCM",
+    price: 9000000,
+    area: 45,
+    bedrooms: 1,
+    bathrooms: 1,
+    owner: "Sky Garden",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=7",
+    date: "2025-02-20",
+    category: "Cho thuê",
+    tags: ["Dịch vụ", "Trung tâm", "Sang trọng"],
+    views: 208,
+    status: "Còn trống",
+    image: "https://picsum.photos/400/300?random=7"
+  },
+  {
+    id: 8,
+    title: "Căn hộ 1 phòng ngủ Sunrise City",
+    description: "Căn hộ thiết kế hiện đại, view đẹp, tiện ích đầy đủ, bảo vệ 24/7",
+    address: "Quận 7, TP.HCM",
+    price: 4200000,
+    area: 50,
+    bedrooms: 1,
+    bathrooms: 1,
+    owner: "Osaka Complex",
+    avatar: "https://xsgames.co/randomusers/avatar.php?g=pixel&key=8",
+    date: "2025-02-15",
+    category: "Cho thuê",
+    tags: ["Hiện đại", "Tiện ích đầy đủ", "View đẹp"],
+    views: 156,
+    status: "Đã cho thuê",
+    image: "https://picsum.photos/400/300?random=8"
   }
-`;
+];
 
-const SearchContainer = styled.div`
-  background: #fde68a;
-  margin-top: 16px;
-  border-radius: 8px;
-  padding: 16px;
-`;
+// Các danh mục căn hộ
+const categories = ["Tất cả", "Cho thuê", "Bán", "Đã cho thuê", "Đang đặt cọc"];
 
-const SearchRow = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-`;
+// Các khu vực
+const areas = ["Tất cả", "Quận 1", "Quận 2", "Quận Bình Thạnh", "Quận 7", "TP. Thủ Đức"];
 
-const SearchInput = styled.input`
-  flex: 1;
-  padding: 8px 12px;
-  border: none;
-  outline: none;
-  border-radius: 4px;
-`;
+// Màu trạng thái
+const statusColors = {
+  "Còn trống": "green",
+  "Đã cho thuê": "red",
+  "Đang đặt cọc": "orange"
+};
 
-const SearchButton = styled.button`
-  background: #dc2626;
-  color: white;
-  padding: 8px 16px;
-  border: none;
-  cursor: pointer;
-  border-radius: 4px;
-  font-weight: 500;
-`;
+const ApartmentList = () => {
+  const [apartments, setApartments] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [searchText, setSearchText] = useState("");
+  const [selectedCategory, setSelectedCategory] = useState("Tất cả");
+  const [selectedArea, setSelectedArea] = useState("Tất cả");
+  const [currentPage, setCurrentPage] = useState(1);
+  const pageSize = 8;
+  
+  // Kích hoạt hook useNavigate để sử dụng điều hướng
+  const navigate = useNavigate();
 
-const FiltersRow = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-top: 16px;
-`;
+  // Giả lập việc lấy dữ liệu từ API
+  useEffect(() => {
+    setTimeout(() => {
+      setApartments(sampleApartments);
+      setLoading(false);
+    }, 1000);
+  }, []);
 
-const Select = styled.select`
-  padding: 8px;
-  border: 1px solid #d1d5db;
-  border-radius: 4px;
-`;
+  // Xử lý lọc căn hộ theo danh mục và khu vực
+  const filteredApartments = apartments.filter(apartment => {
+    const matchSearch = apartment.title.toLowerCase().includes(searchText.toLowerCase()) || 
+                       apartment.description.toLowerCase().includes(searchText.toLowerCase());
+    const matchCategory = selectedCategory === "Tất cả" || apartment.category === selectedCategory;
+    const matchArea = selectedArea === "Tất cả" || apartment.address.includes(selectedArea);
+    
+    return matchSearch && matchCategory && matchArea;
+  });
 
-const MainContent = styled.div`
-  display: grid;
-  grid-template-columns: 3fr 1fr;
-  gap: 16px;
-  margin-top: 16px;
+  // Phân trang
+  const paginatedApartments = filteredApartments.slice(
+    (currentPage - 1) * pageSize,
+    currentPage * pageSize
+  );
 
-  @media (max-width: 768px) {
-    grid-template-columns: 1fr;
-  }
-`;
+  const onSearch = value => {
+    setSearchText(value);
+    setCurrentPage(1);
+  };
 
-const LeftSection = styled.div``;
+  const onCategoryChange = value => {
+    setSelectedCategory(value);
+    setCurrentPage(1);
+  };
 
-const RightSection = styled.div`
-  background: white;
-  padding: 16px;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-`;
+  const onAreaChange = value => {
+    setSelectedArea(value);
+    setCurrentPage(1);
+  };
 
-const ListingColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-`;
+  const formatPrice = (price) => {
+    return new Intl.NumberFormat('vi-VN').format(price) + " VNĐ/tháng";
+  };
+  
+  // Chuyển đến trang chi tiết - đã mở comment và thêm điều hướng thực tế
+  const goToDetails = (id) => {
+    navigate(`/post-detail`);
+    console.log(`Đang chuyển đến trang chi tiết của căn hộ ID: ${id}`);
+  };
 
-/* Link bọc quanh ListingCard, để click vào sẽ sang /apartment-detail */
-const ListingCardLink = styled(Link)`
-  text-decoration: none; 
-  color: inherit;
-`;
-
-const ListingCard = styled.div`
-  display: flex;
-  background: white;
-  padding: 16px;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  gap: 16px;
-`;
-
-const ListingImage = styled.div`
-  width: 120px;
-  height: 120px;
-  background: #e5e7eb;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const ListingInfo = styled.div`
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-`;
-
-const ListingTitle = styled.h2`
-  margin: 0;
-  font-size: 16px;
-  font-weight: bold;
-  color: #111827;
-`;
-
-const ListingText = styled.p`
-  margin: 4px 0;
-  font-size: 14px;
-  color: #6b7280;
-`;
-
-const PriceText = styled.span`
-  color: #f97316;
-  font-weight: 500;
-`;
-
-const MetaRow = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin: 8px 0;
-  font-size: 14px;
-  color: #374151;
-
-  svg {
-    margin-right: 4px;
-  }
-`;
-
-export default function ApartmentList() {
-  const [search, setSearch] = useState("");
-
-  // Vài ví dụ data
-  const dummyApartments = [1, 2, 3];
+  // Xử lý thêm mới
+  const handleAddNew = () => {
+    navigate('/apartment/new');
+  };
 
   return (
-    <Container>
-      <Wrapper>
-        <TopBar>
-          <span>Kênh chủ căn hộ</span>
-          <span>Kênh người thuê căn hộ</span>
-        </TopBar>
-
-        <MainNavBar>
-          <Logo>ABMS</Logo>
-          <NavLinks>
-            <NavItem to="/forum">Diễn đàn</NavItem>
-            <NavItem to="/rent">Cho thuê</NavItem>
-            <NavItem to="/buy">Mua bán</NavItem>
-            <Spacer />
-            <NavItem to="/signIn">Đăng nhập</NavItem>
-            <NavItem to="/signUp">Đăng ký</NavItem>
-          </NavLinks>
-        </MainNavBar>
-
-        <SearchContainer>
-          <SearchRow>
-            <FaSearch />
-            <SearchInput
-              placeholder="Tìm kiếm"
-              value={search}
-              onChange={(e) => setSearch(e.target.value)}
-            />
-            <SearchButton>Tìm kiếm</SearchButton>
-          </SearchRow>
-          <FiltersRow>
-            <Select><option>Diện tích</option></Select>
-            <Select><option>Giá tiền</option></Select>
-            <Select><option>Hình thức</option></Select>
-            <Select><option>Số phòng</option></Select>
-            <Select><option>Lọc</option></Select>
-          </FiltersRow>
-        </SearchContainer>
-
-        <MainContent>
-          <LeftSection>
-            <ListingColumn>
-              {dummyApartments.map((item) => (
-                /* Bọc ListingCardLink, link tới /apartment-detail */
-                <ListingCardLink key={item} to="/apartmentDetail">
-                  <ListingCard>
-                    <ListingImage>
-                      <BsHouseDoor size={40} style={{ color: "#9ca3af" }} />
-                    </ListingImage>
-                    <ListingInfo>
-                      <div>
-                        <ListingTitle>Căn hộ view đẹp</ListingTitle>
-                        <ListingText>
-                          <PriceText>8,5 tr / tháng</PriceText>
-                        </ListingText>
-                        <ListingText>86m2</ListingText>
-                        <MetaRow>
-                          <div><FaBed /> 2</div>
-                          <div><FaBath /> 2</div>
-                          <div><FaCar /> 1</div>
-                        </MetaRow>
-                        <ListingText>Hòa Hải Ngũ Hành Sơn</ListingText>
-                      </div>
-                      <ListingText style={{ fontSize: "12px", color: "#9ca3af" }}>
-                        2 giờ trước
-                      </ListingText>
-                    </ListingInfo>
-                  </ListingCard>
-                </ListingCardLink>
+    <div style={{ padding: "20px" }}>
+      <Card 
+        title={
+          <Title level={3} style={{ textAlign: "center", marginBottom: 0 }}>
+            <HomeOutlined style={{ marginRight: 8 }} />
+            DANH SÁCH CĂN HỘ CHUNG CƯ
+          </Title>
+        } 
+      >
+        <Space style={{ marginBottom: 20 }} size="large" wrap>
+          <Search
+            placeholder="Tìm kiếm căn hộ"
+            onSearch={onSearch}
+            style={{ width: 300 }}
+            prefix={<SearchOutlined />}
+            allowClear
+          />
+          
+          <Space>
+            <FilterOutlined />
+            <Select 
+              defaultValue="Tất cả" 
+              style={{ width: 150 }}
+              onChange={onCategoryChange}
+            >
+              {categories.map(cat => (
+                <Option key={cat} value={cat}>{cat}</Option>
               ))}
-            </ListingColumn>
-          </LeftSection>
+            </Select>
+          </Space>
 
-          <RightSection>
-            <h3>Lọc theo giá tiền</h3>
-            <p>Dưới 3 triệu</p>
-            <p>3 - 4 triệu</p>
-            <p>4 - 5 triệu</p>
-            <p>5 - 6 triệu</p>
-          </RightSection>
-        </MainContent>
-      </Wrapper>
-    </Container>
+          <Space>
+            <EnvironmentOutlined />
+            <Select 
+              defaultValue="Tất cả" 
+              style={{ width: 150 }}
+              onChange={onAreaChange}
+            >
+              {areas.map(area => (
+                <Option key={area} value={area}>{area}</Option>
+              ))}
+            </Select>
+          </Space>
+          
+          <Button 
+            type="primary" 
+            icon={<HomeOutlined />}
+            onClick={handleAddNew}
+          >
+            Thêm căn hộ mới
+          </Button>
+        </Space>
+
+        <Row gutter={[16, 16]}>
+          {loading ? (
+            // Hiển thị skeleton loading
+            Array(4).fill(null).map((_, index) => (
+              <Col xs={24} sm={12} md={8} lg={6} key={`loading-${index}`}>
+                <Card>
+                  <Skeleton.Image style={{ width: '100%', height: 200 }} active />
+                  <Skeleton active paragraph={{ rows: 3 }} />
+                </Card>
+              </Col>
+            ))
+          ) : (
+            // Hiển thị danh sách căn hộ dạng card
+            paginatedApartments.map(apartment => (
+              <Col xs={24} sm={12} md={8} lg={6} key={apartment.id}>
+                <Card
+                  hoverable
+                  cover={
+                    <div style={{ position: 'relative', height: 200, overflow: 'hidden' }}>
+                      <img 
+                        alt={apartment.title}
+                        src={apartment.image}
+                        style={{ width: '100%', height: '100%', objectFit: 'cover' }}
+                        onClick={() => goToDetails(apartment.id)}
+                      />
+                      <Badge 
+                        count={apartment.status}
+                        style={{ 
+                          position: 'absolute', 
+                          top: 10, 
+                          right: 10,
+                          backgroundColor: statusColors[apartment.status]
+                        }}
+                      />
+                      <div 
+                        style={{ 
+                          position: 'absolute',
+                          bottom: 0,
+                          left: 0,
+                          right: 0,
+                          background: 'rgba(0,0,0,0.6)',
+                          padding: '5px 10px',
+                          color: 'white'
+                        }}
+                      >
+                        <Text style={{ color: 'white', fontWeight: 'bold' }}>{apartment.owner}</Text>
+                      </div>
+                    </div>
+                  }
+                  onClick={() => goToDetails(apartment.id)}
+                  actions={[
+                    <Space><AreaChartOutlined key="area" />{`${apartment.area} m²`}</Space>,
+                    <Space><UserOutlined key="rooms" />{`${apartment.bedrooms} PN, ${apartment.bathrooms} VS`}</Space>,
+                    <Space><EyeOutlined key="view" />{apartment.views}</Space>
+                  ]}
+                >
+                  <Card.Meta
+                    title={<Tooltip title={apartment.title}>{apartment.title.length > 28 ? `${apartment.title.substring(0, 28)}...` : apartment.title}</Tooltip>}
+                    description={
+                      <>
+                        <Paragraph ellipsis={{ rows: 2 }} style={{ height: 40 }}>
+                          {apartment.description}
+                        </Paragraph>
+                        <Space direction="vertical" style={{ width: '100%' }}>
+                          <div>
+                            <EnvironmentOutlined style={{ marginRight: 5 }} />
+                            <Text type="secondary">{apartment.address}</Text>
+                          </div>
+                          <div>
+                            <DollarOutlined style={{ marginRight: 5 }} />
+                            <Text strong>{formatPrice(apartment.price)}</Text>
+                          </div>
+                          <div>
+                            <Space wrap>
+                              {apartment.tags.slice(0, 2).map(tag => (
+                                <Tag color="blue" key={tag}>
+                                  {tag}
+                                </Tag>
+                              ))}
+                              {apartment.tags.length > 2 && <Tag>+{apartment.tags.length - 2}</Tag>}
+                            </Space>
+                          </div>
+                          <Button type="link" style={{ padding: 0 }} onClick={(e) => {
+                            e.stopPropagation();
+                            goToDetails(apartment.id);
+                          }}>
+                            Xem chi tiết <ArrowRightOutlined />
+                          </Button>
+                        </Space>
+                      </>
+                    }
+                  />
+                </Card>
+              </Col>
+            ))
+          )}
+        </Row>
+
+        <div style={{ textAlign: 'center', marginTop: 24 }}>
+          <Pagination
+            current={currentPage}
+            pageSize={pageSize}
+            total={filteredApartments.length}
+            onChange={page => setCurrentPage(page)}
+            showSizeChanger={false}
+          />
+        </div>
+      </Card>
+    </div>
   );
-}
+};
+
+export default ApartmentList;
